requirements.txt
flask==1.1.2
gunicorn==20.1.0
PyYAML==6.0
redis==3.5.3
rq==1.10.0
